+++ evebs/models/universe_system.py
@@ -6,8 +6,8 @@ class UniverseSystem(db.Model):
     __tablename__ = 'universe_systems'
 
     id = db.Column(db.BigInteger, primary_key=True)
-    cpp_system_id = db.Column(db.Integer, nullable=False, unique=True)
-    cpp_star_id = db.Column(db.Integer, nullable=False)
+    cpp_system_id = db.Column(db.BigInteger, nullable=False, unique=True)
+    cpp_star_id = db.Column(db.BigInteger, nullable=False)
 
     universe_constellation_id = db.Column(db.BigInteger, db.ForeignKey('universe_constellations.id'))
